requirements.txt
amqp==5.0.6
asgiref==3.4.1
backports.csv==1.0.7
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cheroot==8.5.2
CherryPy==18.6.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-filter==21.1
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
feedparser==6.0.8
future==0.18.2
google-api-core==1.31.2
google-auth==1.35.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.0
google-cloud-texttospeech==2.5.2
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.3.0
jaraco.text==3.5.1
Jinja2==3.0.3
joblib==1.0.1
kombu==5.1.0
lxml==4.6.3
MarkupSafe==2.0.1
more-itertools==8.8.0
mysqlclient==2.1.0
nltk==3.6.2
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
paramiko==2.7.2
Pattern==2.6
pdfminer.six==20201018
portend==2.7.1
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-docx==0.8.11
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
pywin32==302
redis==3.5.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
sgmllib3k==1.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
storage==0.0.4.3
tempora==4.1.1
text-unidecode==1.3
tqdm==4.62.1
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zc.lockfile==2.0
